requirements.txt
torch
torchvision
scikit-learn
pandas
matplotlib
numpy
pyyaml
pyreadr
